Fixes a problem when using attributes in `sch:value-of`
GN uses `xsl:copy-of` to write the `sch:value-of` sentencences in a schematron.
When writing the value of an attribute, GN tries to add the node type attribute instead
of a text, causing an exception:
net.sf.saxon.trans.XPathException: An attribute node (xlink:href) cannot be created after the children of the containing element

In this commit `normalize-space` is use to convert from attribute type node to text type
node.